fix (CropModel) picture orientation issue
